Add seeder and modify the migration.

+++ albums_migration.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS albums (
   artist VARCHAR(512) NOT NULL,
   name VARCHAR(512) NOT NULL,
   release_date INT NOT NULL,
-  sales FLOAT(5, 2) NOT NULL,
+  sales FLOAT NOT NULL,
   genre CHAR(255) NOT NULL,
   PRIMARY KEY (id)
 );
